Move functions from traitsui.tests._tools to traitsui.testing
