+++ traitsui/testing/tester/ui_wrapper.py
@@ -16,8 +16,10 @@ from traitsui.testing.tester.exceptions import (
     LocationNotSupported,
 )
 from traitsui.testing.tester import locator
-from traitsui.tests._tools import (
+from traitsui.testing.gui import (
     process_cascade_events as _process_cascade_events,
+)
+from traitsui.testing.exception_handling import (
     reraise_exceptions as _reraise_exceptions,
 )
