+++ traitsui/testing/tests/test_gui.py
@@ -9,21 +9,19 @@
 #  Thanks for using Enthought open source!
 #
 
-""" Tests for traitsui.tests._tools """
+""" Tests for traitsui.testing.gui """
 
 import os
 import unittest
 
 from pyface.api import GUI
 
-from traits.api import HasTraits, Int
 from traitsui.tests._tools import (
     is_qt,
     is_wx,
     is_mac_os,
     process_cascade_events,
     requires_toolkit,
-    reraise_exceptions,
     ToolkitName,
 )
 
@@ -180,54 +178,3 @@ class TestProcessEventsRepeated(unittest.TestCase):
 
         # then
         self.assertEqual(wx_handler.n_events, max_n_events)
-
-
-class TestExceptionHandling(unittest.TestCase):
-
-    @requires_toolkit([ToolkitName.qt, ToolkitName.wx])
-    def test_error_from_gui_captured_and_raise(self):
-
-        def raise_error_1():
-            raise ZeroDivisionError()
-
-        def raise_error_2():
-            raise IndexError()
-
-        # without the context manager:
-        #   - with Qt5, the test run will be aborted prematurely.
-        #   - with Qt4, the traceback is printed and the test passes.
-        #   - with Wx, the traceback is printed and the test passes.
-        # With the context manager, the exception is always reraised.
-        gui = GUI()
-        with self.assertRaises(RuntimeError) as exception_context, \
-                self.assertLogs("traitsui") as watcher:
-            with reraise_exceptions():
-                gui.invoke_later(raise_error_1)
-                gui.invoke_later(raise_error_2)
-                gui.process_events()
-
-        error_msg = str(exception_context.exception)
-        self.assertIn("ZeroDivisionError", error_msg)
-        self.assertIn("IndexError", error_msg)
-        log_content1, log_content2 = watcher.output
-        self.assertIn("ZeroDivisionError", log_content1)
-        self.assertIn("IndexError", log_content2)
-
-    def test_error_from_trait_change_captured(self):
-
-        class Foo(HasTraits):
-            value = Int()
-
-            def _value_changed(self):
-                raise ZeroDivisionError()
-
-        obj = Foo()
-        with self.assertRaises(RuntimeError) as exception_context, \
-                self.assertLogs("traitsui") as watcher:
-            with reraise_exceptions():
-                obj.value = 2
-
-        error_msg = str(exception_context.exception)
-        self.assertIn("ZeroDivisionError", error_msg)
-        log_content, = watcher.output
-        self.assertIn("ZeroDivisionError", log_content)
